pyproject.toml
[project]
name = "llm-tech-docs-rag"
version = "0.1.0"
description = "A RAG system for technical documentation using LLMs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.10.0",
    "llama-index-llms-openai>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-embeddings-huggingface>=0.6.0",
    "llama-index-readers-file>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "markdown>=3.4.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "streamlit>=1.48.1",
    "llama-index-llms-ollama>=0.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "C",   # flake8-comprehensions
  "B",   # flake8-bugbear
  "Q",   # flake8-quotes
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "mypy>=1.5.0",
    "ruff>=0.0.285",
    "pre-commit>=3.0.0",
]
